test: Actor 테스트 코드를 joka-api에서 domain-actor 모듈로 이동
