--- packages/domain-actor/test/temp.test.ts
@@ -1,13 +1,0 @@
-describe('temp', () => {
-  it('1 + 1 = 2', async () => {
-    // given
-    const a = 1;
-    const b = 1;
-
-    // when
-    const result = a + b;
-
-    // then
-    expect(result).toBe(2);
-  });
-});
